pyproject.toml
[project]
name = "amiablog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]
